requirements.txt
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
transformers==4.57.3
diffusers==0.35.2
accelerate==1.12.0
timm==1.0.22
sentencepiece==0.2.1
datasets==4.4.1
huggingface_hub==0.36.0
pillow==12.0.0
scikit-learn==1.7.2
pandas==2.3.3
tqdm==4.67.1
requests==2.32.5
pyarrow==22.0.0
flask==3.1.0
joblib==1.4.2
